+++ CHANGELOG.md
@@ -10,10 +10,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 -   [🚀] Added option to search for a location inside location block (@nehnehneh)
+-   [🚀] Supporting alternate syntax using [latitude, longitude] array (#18)
+-   [🚀] Allow reverse order of coordinates (e.g. if copied from Google Maps) via setting (#21)
 
 ### Changed
 
 -   [⚙️] ci: added step to check commit messages on build workflow
+
+### Fixed
+
+-   [🐛] Fixed issue with custom marker icon not being displayed correctly
 
 ## [1.2.0] - 2024-07-26